ocr/requirements.txt
astroid==3.1.0
dill==0.3.8
dnspython==2.6.1
isort==5.13.2
jedi==0.19.1
mccabe==0.7.0
packaging==24.0
parso==0.8.4
pika==1.3.2
pillow==10.3.0
platformdirs==4.2.0
pylint==3.1.0
pymongo==4.6.3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
six==1.16.0
tesseract==0.1.3
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
